Refactor: Reorganizing the project tree.
* Removed boilerplate code.
* Tried to organize it a bit more.
* Logics are no more in mod.rs, they were moved to common.rs instead.
* Flake.nix allows cross compilation from a devShell or nix build
.#target
