+++ .cargo/config.toml
@@ -1,2 +1,2 @@
-# [build]
-# target = "x86_64-pc-windows-gnu"
+[build]
+target = "x86_64-pc-windows-gnu"
