+++ src/platform/linux/mod.rs
@@ -1,2 +1,4 @@
+pub mod common;
 pub mod process;
-pub mod util;
+pub mod wayland;
+pub mod x11;
