+++ src/platform/linux/x11.rs
@@ -1,20 +1,15 @@
-use std::env;
-use std::fs;
-use std::io::Write;
-use std::path::Path;
-use std::process::Command;
-use std::time::Duration;
-
-use serde::Deserialize;
 use x11rb::connection::Connection;
 use x11rb::protocol::screensaver;
 use x11rb::protocol::xproto::*;
 use x11rb::rust_connection::RustConnection;
 
-use tracing::*;
+use crate::platform::linux::common::*;
+
+use tokio::time::Duration;
 
 // https://www.reddit.com/r/rust/comments/f7yrle/get_information_about_current_w_xorg/
 // Returns name and class of the focused window in that order.
+/// Find focused window in x11 environment.
 pub fn get_focused_window() -> Result<(String, String), Box<dyn std::error::Error>> {
     // Set up our state
     // TODO: This will fucking panic for no reason
@@ -49,15 +44,12 @@ pub fn get_focused_window() -> Result<(String, String), Box<dyn std::error::Erro
     Ok((name, class))
 }
 
-pub fn get_screen_dpi() -> Result<f64, Box<dyn std::error::Error>> {
-    // Open connection to the X server
+// TODO:
+pub fn _get_screen_dpi() -> Result<f64, Box<dyn std::error::Error>> {
     let (conn, screen_num) = RustConnection::connect(None)?;
 
-    // Get the setup and screen information
     let setup = conn.setup();
     let screen = &setup.roots[screen_num];
-
-    // Get screen dimensions in pixels and millimeters
     let width_px = screen.width_in_pixels as f64;
     let height_px = screen.height_in_pixels as f64;
 
@@ -68,41 +60,8 @@ pub fn get_screen_dpi() -> Result<f64, Box<dyn std::error::Error>> {
     let dpi_x = (width_px / width_mm) * 25.4; // 25.4 mm in an inch
     let dpi_y = (height_px / height_mm) * 25.4;
 
-    // Use the average DPI of the x and y dimensions
     let average_dpi = (dpi_x + dpi_y) / 2.0;
-
     Ok(average_dpi)
-}
-
-#[derive(Debug, Clone, Copy, Deserialize)]
-pub struct MouseSettings {
-    pub threshold: u16,
-    pub acceleration_denominator: u16,
-    pub acceleration_numerator: u16,
-    pub dpi: u32,
-}
-
-// Default values from ArchLinux, didn't check for other OS's.
-impl Default for MouseSettings {
-    fn default() -> Self {
-        MouseSettings {
-            threshold: 4,
-            acceleration_numerator: 2,
-            acceleration_denominator: 1,
-            dpi: 800,
-        }
-    }
-}
-
-impl MouseSettings {
-    pub fn noacc_default() -> Self {
-        MouseSettings {
-            acceleration_numerator: 1,
-            acceleration_denominator: 1,
-            threshold: 0,
-            dpi: 800,
-        }
-    }
 }
 
 pub fn get_mouse_settings() -> Result<MouseSettings, Box<dyn std::error::Error>> {
@@ -131,120 +90,13 @@ pub fn get_mouse_settings() -> Result<MouseSettings, Box<dyn std::error::Error>>
     Ok(s)
 }
 
-// TODO:
-// Add flags from somewhere i dont know where yet
-pub fn configure_startup(
-    should_enable: bool,
-    is_enable: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let service_name = "life-monitor.service";
-    let user_service_dir = Path::new(&env::var("HOME")?).join(".config/systemd/user");
-    fs::create_dir_all(&user_service_dir)?;
-    let service_path = user_service_dir.join(service_name);
-    let current_exe = env::current_exe()?;
-
-    if should_enable {
-        if is_enable {
-            info!("Startup is already enabled!");
-            return Ok(());
-        }
-        info!("Creating service for life-monitor");
-
-        let service_content = format!(
-            "[Unit]\n\
-            Description=Life Monitor Service\n\
-            After=display-manager.service\n\
-            Wants=graphical-session.target multi-user.target\n\
-            \n\
-            [Service]\n\
-            Type=simple\n\
-            Environment=DISPLAY=:0\n\
-            Environment=XAUTHORITY=/home/{}/.Xauthority\n\
-            ExecStart={}\n\
-            Restart=always\n\
-            ExecStartPre=/bin/sh -c 'until [ -n \"$DISPLAY\" ] && xset q; do sleep 1; done'
-            \n\
-            [Install]\n\
-            WantedBy=graphical-session.target multi-user.target\n",
-            env::var("USER")?,
-            current_exe.to_str().unwrap()
-        );
-
-        // Write service file
-        let mut file = fs::OpenOptions::new()
-            .write(true)
-            .create(true)
-            .truncate(true)
-            .open(&service_path)?;
-        file.write_all(service_content.as_bytes())?;
-
-        // Enable and start the user service
-        Command::new("systemctl")
-            .args(["--user", "daemon-reload"])
-            .status()?;
-        Command::new("systemctl")
-            .args(["--user", "enable", service_name])
-            .status()?;
-        Command::new("systemctl")
-            .args(["--user", "start", service_name])
-            .status()?;
-        info!("Created and enabled user systemd service: {}", service_name);
-    } else {
-        // Disable and stop the user service
-        Command::new("systemctl")
-            .args(["--user", "stop", service_name])
-            .status()?;
-        Command::new("systemctl")
-            .args(["--user", "disable", service_name])
-            .status()?;
-
-        // Remove service file if it exists
-        if service_path.exists() {
-            fs::remove_file(&service_path)?;
-        }
-        Command::new("systemctl")
-            .args(["--user", "daemon-reload"])
-            .status()?;
-        info!("Removed user systemd service: {}", service_name);
-    }
-    Ok(())
-}
-
-pub fn check_startup_status() -> Result<bool, Box<dyn std::error::Error>> {
-    let service_name = "life-monitor.service";
-
-    // Check if service is enabled
-    let status = Command::new("systemctl")
-        .args(["--user", "is-enabled", service_name])
-        .output()?;
-
-    // Also check if the service file exists
-    let user_service_dir = Path::new(&env::var("HOME")?).join(".config/systemd/user");
-    let service_path = user_service_dir.join(service_name);
-
-    let is_enabled = status.status.success() && service_path.exists();
-
-    info!(
-        "Startup status on Linux is {}.",
-        if is_enabled { "Enabled" } else { "Disabled" }
-    );
-
-    Ok(is_enabled)
-}
-
 pub fn get_idle_time() -> Result<Duration, Box<dyn std::error::Error>> {
-    // Connect to the X server
     let (conn, screen_num) = RustConnection::connect(None)?;
-    // Get the root window of the default screen
     let screen = &conn.setup().roots[screen_num];
-
-    // Query the XScreenSaver extension for idle time
     let reply = screensaver::query_info(&conn, screen.root)?.reply()?;
 
-    // Extract idle time in milliseconds
     let idle_time_ms = reply.ms_since_user_input;
 
-    // Convert to `Duration`
     Ok(Duration::from_millis(idle_time_ms as u64))
 }
