+++ src/platform/windows/mod.rs
@@ -1,3 +1,3 @@
+pub mod common;
 pub mod process;
 pub mod systray;
-pub mod util;
