+++ src/storage/mod.rs
@@ -1,2 +1,3 @@
 pub mod api;
+pub mod backend;
 pub mod localdb;
